feat: :safety_vest: add validation in email
Add the email is not valid in controller postUsers function
